refactor: move to own package: promotion
