+++ card/example_test.go
@@ -2,11 +2,9 @@ package card_test
 
 import (
 	"fmt"
-	"log"
-	"time"
-
 	"github.com/xendit/xendit-go"
 	"github.com/xendit/xendit-go/card"
+	"log"
 )
 
 func ExampleCreateCharge() {
@@ -91,34 +89,4 @@ func ExampleReverseAuthorization() {
 	}
 
 	fmt.Printf("reversed authorization: %+v\n", reverseAuthorizationResp)
-}
-
-func ExampleCreatePromotion() {
-	xendit.Opt.SecretKey = "examplesecretkey"
-
-	startTime := time.Now().Add(time.Hour)
-	endTime := startTime.Add(time.Hour)
-
-	createPromotionData := card.CreatePromotionParams{
-		ReferenceID: "BRI_20_JAN",
-		Description: "20% discount applied for all BRI cards",
-		BinList: []string{
-			"400000",
-			"460000",
-		},
-		DiscountPercent:   20,
-		Currency:          "IDR",
-		StartTime:         &startTime,
-		EndTime:           &endTime,
-		ChannelCode:       "BRI",
-		MinOriginalAmount: 25000,
-		MaxDiscountAmount: 5000,
-	}
-
-	promotionResp, err := card.CreatePromotion(&createPromotionData)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	fmt.Printf("created promotion: %+v\n", promotionResp)
 }
