+++ integration_test/main.go
@@ -11,17 +11,13 @@ func main() {
 	xendit.Opt.SecretKey = os.Getenv("SECRET_KEY")
 
 	wg := sync.WaitGroup{}
-	wg.Add(10)
+	wg.Add(11)
 	go func() {
 		balanceTest()
 		wg.Done()
 	}()
 	go func() {
 		cardTest()
-		wg.Done()
-	}()
-	go func() {
-		cardPromotionTest()
 		wg.Done()
 	}()
 	go func() {
@@ -54,6 +50,10 @@ func main() {
 	}()
 	go func() {
 		virtualaccountTest()
+		wg.Done()
+	}()
+	go func() {
+		promotionTest()
 		wg.Done()
 	}()
